// apps/api/gql/generated-schema.ts
export type Maybe<T> = T | null;
// ...
export interface Scalars {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    AmountHumanReadable: string;
    BigDecimal: string;
    BigInt: string;
    Bytes: string;
    Date: Date;
    GqlBigNumber: string;
    JSON: any;
}
// ...
export type GqlChain =
    | 'ARBITRUM'
    | 'AVALANCHE'
    | 'BASE'
    | 'BERACHAIN'
    | 'FANTOM'
    | 'FRAXTAL'
    | 'GNOSIS'
    | 'HYPEREVM'
    | 'HYPEREVM_TESTNET'
    | 'MAINNET'
    | 'MODE'
    | 'OPTIMISM'
    | 'POLYGON'
    | 'SEPOLIA'
    | 'SONIC'
    | 'ZKEVM';

export interface GqlFeaturePoolGroupItemExternalLink {
    __typename?: 'GqlFeaturePoolGroupItemExternalLink';
    buttonText: Scalars['String'];
    buttonUrl: Scalars['String'];
    id: Scalars['ID'];
    image: Scalars['String'];
}

export interface GqlHistoricalTokenPrice {
    __typename?: 'GqlHistoricalTokenPrice';
    address: Scalars['String'];
    chain: GqlChain;
    prices: Array<GqlHistoricalTokenPriceEntry>;
}

export interface GqlHistoricalTokenPriceEntry {
    __typename?: 'GqlHistoricalTokenPriceEntry';
    price: Scalars['Float'];
    timestamp: Scalars['String'];
    updatedAt: Scalars['Int'];
    updatedBy?: Maybe<Scalars['String']>;
}